fix: 총 수익이 Integer범위를 벗어나는 경우 수익률을 -로 표시하던 문제 수정

+++ src/main/java/lotto/LottoNumberMatcher.java
@@ -48,9 +48,9 @@ public class LottoNumberMatcher {
         return Math.round(earnRate * 100) / 100.0;
     }
 
-    private Integer calculateTotalEarning(List<WinRank> winningWinRanks) {
+    private Long calculateTotalEarning(List<WinRank> winningWinRanks) {
         return winningWinRanks.stream()
-                .mapToInt(LottoNumberMatcher::winPriceFromRank)
+                .mapToLong(LottoNumberMatcher::winPriceFromRank)
                 .sum();
     }
